Create PointQuadtree.java
Quadtrees are an efficient data-structure used to represent multi-dimensional spatial data. Quadtrees
come in many flavours, we shall be looking at one of them, namely Point Quadtrees, in this assignment to represent points on a 2-dimensional cartesian plane.
A Point Quadtree contains a point in the cartesian plane at every node. Further, a point (x0, y0)
in it has 4 children, each representing a quadrant obtained by lines x = x0 and y = y0. (See
Figure below). Note that, the structure of the Quadtree is dependent on the order of insertion.